pyproject.toml
[project]
name = "mcp-invest"
version = "0.1.0"
description = "MCP Python server for invest SDK"
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.7.1",
    "tinkoff-investments>=0.2.0b111",
]
